Finished first implementation of SPLIT player action

+++ cogs/blackjack.py
@@ -21,7 +21,12 @@ usersCol = myDB[db["USERS_COL"]]
 blackjackGameCol = myDB["BLACKJACK_GAME"]
 blackjackUserCol = myDB["BLACKJACK_USER"]
 
-cardSuits = ["♦︎", "♣︎", "♥︎", "♠︎"]
+adminRole = global_json["ROLES"]["ADMIN_ROLE"]
+
+playerColours = [0xfc0377, 0x03fceb, 0xe7fc03, 0xfc7703]
+
+cardSuits = [":diamond_suit:", ":club_suit:", ":heart_suit:", ":spade_suit:"]
+
 cardRanks = {
     "A": 1, 
     "2": 2,
@@ -36,6 +41,22 @@ cardRanks = {
     "J": 10,
     "Q": 10,
     "K": 10
+}
+
+cardRanksEmoji = {
+    "A": ":a:", 
+    "2": ":number_2:",
+    "3": ":number_3:",
+    "4": ":number_4:",
+    "5": ":number_5:",
+    "6": ":number_6:",
+    "7": ":number_7:",
+    "8": ":number_8:",
+    "9": ":number_9:",
+    "10": ":number_10:",
+    "J": ":jack_o_lantern:",
+    "Q": ":princess:",
+    "K": ":crown:"
 }
 
 class Card:
@@ -108,8 +129,8 @@ def draw_card(playDeck):
 def print_hand(hand):
     handStr = "Your current hand:\n"
     for card in hand:
-        handStr += card["rank"] + " of " + card["suit"] + "\n"
-    handStr += "\nTotal value = " + str(evaluate_hand(hand))
+        handStr += cardRanksEmoji[card["rank"]] + " of " + card["suit"] + "\n"
+    handStr += "\nTotal value = **" + str(evaluate_hand(hand)) + "**"
     return handStr
 
 def print_house_hand(hand, numHidden):
@@ -117,12 +138,12 @@ def print_house_hand(hand, numHidden):
     value = 0
     for card in hand:
         if numHidden > 0:
-            handStr += "Hidden card\n"
+            handStr += "[redacted]\n"
             numHidden -= 1
             continue
-        handStr += card["rank"] + " of " + card["suit"] + "\n"
+        handStr += cardRanksEmoji[card["rank"]] + " of " + card["suit"] + "\n"
         value += evaluate_hand([card])
-    handStr += "\nTotal value = " + str(value)
+    handStr += "\nTotal value = **" + str(value) + "**"
     return handStr
 
 async def update_wallet(ctx: discord.ApplicationContext, amount: int, remove: bool):
@@ -132,21 +153,59 @@ async def update_wallet(ctx: discord.ApplicationContext, amount: int, remove: bo
         await ctx.respond("OOPS! This user isn't in the database! Notify bot admin!", ephemeral=True)
         return -1
 
+    #When removing coins from wallet
     if remove:
         if(userCheck["coins"] < amount): 
             await ctx.respond("You seem to be too poor right now. Maybe apply for a small loan of a million <:beets:1245409413284499587>?", ephemeral=True)
             return -1
             
-        #Update wallet before game
         removeCoins = 0 - amount
         myQuery= {"member_id": ctx.author.id, "guild_id": ctx.guild.id}
         newValues = {'$inc': {'coins': int(removeCoins), 'coins_bet': int(amount)}}
         usersCol.update_one(myQuery, newValues)
 
+    # When adding coins to wallet
     else:
         myQuery= {"member_id": ctx.author.id, "guild_id": ctx.guild.id}
         newValues = {'$inc': {'coins': int(amount), 'earned_bet': int(amount), 'total_earned': int(amount)}}
         usersCol.update_one(myQuery, newValues)
+
+async def calculate_player_reward(ctx: discord.ApplicationContext, player, houseValue):
+    playerHands = []
+    playerHands.append(player["hand"])
+    if player["split"]:
+        playerHands.append(player["hand2"])
+
+    for hand in playerHands:
+        playerValue = evaluate_hand(hand)
+        bet = player["bet"]
+                
+        if playerValue > 21:
+            await ctx.respond("[Blackjack] <@" + str(player["member_id"]) + 
+                              "> busted! You lost " + str(bet) + 
+                              "<:beets:1245409413284499587>!")
+                    
+        elif houseValue > 21 or playerValue > houseValue:
+            if playerValue == 21:
+                await ctx.respond("[Blackjack] <@" + str(player["member_id"]) + 
+                                  "> had a **Blackjack**! You win **" + str(math.floor(bet*2.5)) + 
+                                  "**<:beets:1245409413284499587>!")
+                await update_wallet(ctx, math.floor(bet*2.5), False)
+
+            await ctx.respond("[Blackjack] <@" + str(player["member_id"]) + 
+                              "> won **" + str(math.floor(bet*2)) + 
+                              "**<:beets:1245409413284499587>!")
+            await update_wallet(ctx, math.floor(bet*2), False)
+                    
+        elif playerValue == houseValue:
+            await ctx.respond("[Blackjack] <@" + str(player["member_id"]) + 
+                              "> ties with the house! You got back your " + 
+                              str(bet) + "<:beets:1245409413284499587>!")
+            await update_wallet(ctx, bet, False)
+                    
+        else:
+            await ctx.respond("[Blackjack] <@" + str(player["member_id"]) + 
+                              "> loses " + str(bet) + "<:beets:1245409413284499587>!")
 
 class Blackjack(commands.Cog):
 
@@ -165,8 +224,7 @@ class Blackjack(commands.Cog):
             newGame = {
                 "guild_id": ctx.guild.id,
                 "author_id": ctx.author.id,
-                "running": True,
-                "rolling": False
+                "running": True
             }
             blackjackGameCol.insert_one(newGame)
             self.queue.clear()
@@ -232,23 +290,27 @@ class Blackjack(commands.Cog):
 
             blackjackUserCol.update_one({"member_id": player, "guild_id": ctx.guild.id}, 
                                         {"$set": {"hand": playerHand}})
-
+        
         # PLAYER'S TURN (BY SEQUENCE OF JOINING)
         while len(self.queue) > 0:
             player = self.queue.pop()
-            blackjackUserCol.update_one({"member_id": player, "guild_id": ctx.guild.id}, 
-                                        {"$set": {"playing": True}})
+            playerColor = playerColours[len(self.queue)%len(playerColours)]
             
             # Check if player has blackjack
-            playerHand = blackjackUserCol.find_one({"member_id": player, "guild_id": ctx.guild.id},{"_id": 0, "hand": 1})["hand"]
+            query = blackjackUserCol.find_one({"member_id": player, "guild_id": ctx.guild.id},
+                                              {"_id": 0, "hand": 1, "hand2": 1, "split": 1})
+            if query["split"]:
+                playerHand = query["hand2"]
+                split = True
+            else:
+                playerHand = query["hand"]
+                split = False
 
             if evaluate_hand(playerHand) == 21:
-                await ctx.respond("[Blackjack] <@" + str(player) + "> has a Blackjack! ", ephemeral=True)
-                blackjackUserCol.update_one({"member_id": player, "guild_id": ctx.guild.id}, 
-                                        {"$set": {"playing": False}})
+                await ctx.respond("[Blackjack] <@" + str(player) + "> has a Blackjack! Lucky bastard", ephemeral=True)
                 continue
             
-            timeout = 10
+            timeout = 30
             end_action_time = timeout + math.floor(time.time())
             playerHandDesc = print_hand(playerHand)
             houseHandDesc = print_house_hand(houseHand,1)
@@ -258,7 +320,7 @@ class Blackjack(commands.Cog):
                                   playerHandDesc + "\n\n" + houseHandDesc + 
                                   "\n\nType your action: `hit`, `stand`, `double`, or `split`.\n\n\
                                   You have <t:" + str(end_action_time) + ":R> to make your move!",
-                                  colour=0x009900,
+                                  colour=playerColor,
                                   timestamp=datetime.now())
 
             await ctx.respond(embed=embed, allowed_mentions=discord.AllowedMentions())
@@ -285,16 +347,14 @@ class Blackjack(commands.Cog):
                                               playerHandDesc + "\n\n" + houseHandDesc + 
                                               "\n\nType your action: `hit`, `stand`, `double`, or `split`.\n\n\
                                               You have <t:" + str(end_action_time) + ":R> to make your move!",
-                                              colour=0x009900,
+                                              colour=playerColor,
                                               timestamp=datetime.now())
                         await ctx.respond(embed=embed, allowed_mentions=discord.AllowedMentions())
                     
                     # STAND
                     elif message == "stand" or message == "st":
                         flag = 1
-                        blackjackUserCol.update_one({"member_id": player, "guild_id": ctx.guild.id}, 
-                                                    {"$set": {"playing": False}})
-
+                    
                     # DOUBLE    
                     elif message == "double" or message == "d":
                         #Change bet to double if possible
@@ -313,36 +373,72 @@ class Blackjack(commands.Cog):
                         houseHandDesc = print_house_hand(houseHand,1)
                         embed = discord.Embed(title="[Blackjack] Player used DOUBLE!",
                                               description="Player - <@" + str(ctx.author.id) + ">\n\n" + 
-                                              playerHandDesc + "\n\n" + houseHandDesc,
-                                              colour=0x009900,
+                                              playerHandDesc + "\n\n" + houseHandDesc +
+                                              "\n\nType your action: `hit`, `stand`, `double`, or `split`.\n\n\
+                                              You have <t:" + str(end_action_time) + ":R> to make your move!",
+                                              colour=playerColor,
                                               timestamp=datetime.now())
                         await ctx.respond(embed=embed, allowed_mentions=discord.AllowedMentions())
-                        
                         flag = 1
-                        blackjackUserCol.update_one({"member_id": player, "guild_id": ctx.guild.id}, 
-                                                    {"$set": {"playing": False}})
 
                     # SPLIT    
                     elif message == "split" or message == "sp":
                         #WHAT TO DO WHEN SPLIT
-                        print("SPLIT")
+                        if playerHand[0]["rank"] != playerHand[1]["rank"]:
+                            await ctx.respond("You cannot split your hand! Your cards are not of the same rank.", ephemeral=True)
+                            continue
+
+                        #Search for needed values
+                        query = blackjackUserCol.find_one({"member_id": player, 
+                                                            "guild_id": ctx.guild.id},
+                                                            {"_id": 0, "bet": 1, "split": 1})
+                        split = query["split"]
+                        if split:
+                            await ctx.respond("You are only allowed to split once. Make sure to spam RIGGED when you see this message.", ephemeral=True)
+                            continue
+
+                        amount = query["bet"]
+                        if await update_wallet(ctx, amount, True) == -1:
+                            await ctx.respond("You cannot split your hand! You don't have enough <:beets:1245409413284499587>!", ephemeral=True)
+                            continue
+
+                        #Draw cards for both hands
+                        playerHand1 = [playerHand[0], draw_card(playDeck).to_dict()]
+                        playerHand2 = [playerHand[1], draw_card(playDeck).to_dict()]
+
+                        myQuery= {"member_id": player, "guild_id": ctx.guild.id}
+                        newValues = {'$set': {'hand': playerHand1, 
+                                              'hand2': playerHand2,
+                                              'split': True}}
+                        blackjackUserCol.update_one(myQuery, newValues)
+                        playerHand = playerHand1
+
+                        #Add player back to queue so they can play their second hand
+                        self.queue.insert(0, player)
+
+                        playerHandDesc = print_hand(playerHand)
+                        houseHandDesc = print_house_hand(houseHand,1)
+                        embed = discord.Embed(title="[Blackjack] Player used SPLIT!",
+                                              description="Player - <@" + str(ctx.author.id) + ">\n\n" + 
+                                              playerHandDesc + "\n\n" + houseHandDesc,
+                                              colour=playerColor,
+                                              timestamp=datetime.now())
+                        await ctx.respond(embed=embed, allowed_mentions=discord.AllowedMentions())
+
                     else:
                         await ctx.respond("Invalid action! Please type `hit`, `stand`, `double`, or `split`.", ephemeral=True)
                         continue
                     
                 except asyncio.TimeoutError:
                     await ctx.respond("[Blackjack] <@" + str(player) + "> took too long to respond! Standing their hand...")
-                    blackjackUserCol.update_one({"member_id": player, "guild_id": ctx.guild.id}, 
-                                            {"$set": {"playing": False}})
                     flag = 1
                 
                 playerScore = evaluate_hand(playerHand)
                 if playerScore >= 21:
                     flag = 1
-                    blackjackUserCol.update_one({"member_id": player, "guild_id": ctx.guild.id},
-                                                {"$set": {"playing": False}})
                     if(playerScore == 21):
-                        await ctx.respond("[Blackjack] <@" + str(player) + "> has a Blackjack! Advancing to next player...")
+                        await ctx.respond("[Blackjack] <@" + str(player) + "> finally has a Blackjack! \
+                                          Not even close :sunglasses:. Advancing to next player...")
                         
                     elif(playerScore > 21):
                         await ctx.respond("[Blackjack] <@" + str(player) + "> busted! Advancing to next player...")
@@ -350,7 +446,10 @@ class Blackjack(commands.Cog):
 
             # Update final hand to database
             myQuery= {"member_id": player, "guild_id": ctx.guild.id}
-            newValues = {'$set': {'hand': playerHand}}
+            if split:
+                newValues = {'$set': {'hand2': playerHand}}
+            else:
+                newValues = {'$set': {'hand': playerHand}}
             blackjackUserCol.update_one(myQuery, newValues)
 
         embed = discord.Embed(title="[Blackjack] House's turn!",
@@ -361,6 +460,7 @@ class Blackjack(commands.Cog):
 
         # HOUSE TURN
         while evaluate_hand(houseHand) < 17:
+            #Suspense...
             await asyncio.sleep(3)
             houseHand.append(draw_card(playDeck).to_dict())
             embed = discord.Embed(title="[Blackjack] House's turn!",
@@ -370,24 +470,12 @@ class Blackjack(commands.Cog):
             await ctx.respond(embed=embed, allowed_mentions=discord.AllowedMentions())
             
         # CHECK WINNERS
+        await asyncio.sleep(1)
         houseValue = evaluate_hand(houseHand)
-        for player in blackjackUserCol.find({"guild_id": ctx.guild.id},{"_id": 0, "member_id": 1, "hand": 1, "bet": 1}):
-            playerHand = player["hand"]
-            playerValue = evaluate_hand(playerHand)
-            bet = player["bet"]
-            if playerValue > 21:
-                await ctx.respond("[Blackjack] <@" + str(player["member_id"]) + "> busted! You lost " + str(bet) + "<:beets:1245409413284499587>!")
-            elif houseValue > 21 or playerValue > houseValue:
-                if playerValue == 21:
-                    await ctx.respond("[Blackjack] <@" + str(player["member_id"]) + "> has a Blackjack! You win " + str(bet*2.5) + "<:beets:1245409413284499587>!")
-                    await update_wallet(ctx, math.floor(bet*2.5), False)
-                await ctx.respond("[Blackjack] <@" + str(player["member_id"]) + "> wins " + str(bet*2) + "<:beets:1245409413284499587>!")
-                await update_wallet(ctx, math.floor(bet*2), False)
-            elif playerValue == houseValue:
-                await ctx.respond("[Blackjack] <@" + str(player["member_id"]) + "> ties with the house! You get back your " + str(bet) + "<:beets:1245409413284499587>!")
-                await update_wallet(ctx, bet, False)
-            else:
-                await ctx.respond("[Blackjack] <@" + str(player["member_id"]) + "> loses " + str(bet) + "<:beets:1245409413284499587>!")
+        for player in blackjackUserCol.find({"guild_id": ctx.guild.id},
+                                            {"_id": 0, "member_id": 1, "hand": 1, 
+                                             "hand2": 1, "bet": 1, "split": 1}):
+            await calculate_player_reward(ctx, player, houseValue)
 
         # Clean up from database
         myQuery = {"guild_id": ctx.guild.id}
@@ -409,40 +497,36 @@ class Blackjack(commands.Cog):
         
         if await update_wallet(ctx, amount, True) == -1:
             return
-    
-        # #Check wallet
-        # userCheck = usersCol.find_one({"member_id": ctx.author.id, "guild_id": ctx.guild.id},{"_id": 0, "coins": 1})
-        # if(userCheck is None): 
-        #     await ctx.respond("OOPS! This user isn't in the database! Notify bot admin!", ephemeral=True)
-        #     return
-
-        # if(userCheck["coins"] < amount): 
-        #     await ctx.respond("You seem to be too poor right now. Maybe apply for a small loan of a million <:beets:1245409413284499587>?", ephemeral=True)
-        #     return
-        
-        # #Update wallet before game
-        # removeCoins = 0 - amount
-        # myQuery= {"member_id": ctx.author.id, "guild_id": ctx.guild.id}
-        # newValues = {'$inc': {'coins': int(removeCoins), 'coins_bet': int(amount)}}
-        # usersCol.update_one(myQuery, newValues)
 
         newPlayer = {
             "member_id": ctx.author.id,
             "guild_id": ctx.guild.id,
             "bet": amount,
-            "playing": False,
-            "hand": []
+            "hand": [],
+            "hand2": [],
+            "split": False
         }
         blackjackUserCol.insert_one(newPlayer)
         self.queue.append(ctx.author.id)
 
         await ctx.respond("[Blackjack] <@" + str(ctx.author.id) + "> entered the table with a **" + str(amount) + "**<:beets:1245409413284499587> bet!")
 
-    @blackjack.command(name="reset", description="Reset Blackjack DB.")
+    @blackjack.command(name="reset", description="[ADMIN] Reset Blackjack table.")
     async def reset(self, ctx: discord.ApplicationContext):
+        role = discord.utils.get(ctx.author.roles, name=adminRole) #Check if user has the correct role
+        if role is None:
+            await ctx.respond("You don't have the necessary role!", ephemeral=True)
+            return
+        
+        for player in blackjackUserCol.find({"guild_id": ctx.guild.id},{"_id": 0, "member_id": 1,
+                                                                        "bet": 1, "split": 1}):
+            await update_wallet(ctx, player["bet"], False)
+            if player["split"]:
+                await update_wallet(ctx, player["bet"], False)
+
         blackjackGameCol.delete_many({})
         blackjackUserCol.delete_many({})
-        await ctx.respond("Blackjack database reset!")
+        await ctx.respond("Blackjack table reset!")
 
 def setup(bot):
     bot.add_cog(Blackjack(bot))
